requirements.txt
google-adk
fastmcp
mcp
litellm
uvicorn
fastapi
